requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
anyio==3.6.2
appdirs==1.4.4
asttokens==2.2.1
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.2
billiard==3.6.4.0
blinker==1.5
blobfile==2.0.1
Brotli==1.0.9
bs4==0.0.1
celery==5.2.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
crcmod==1.7
cryptography==39.0.2
cycler==0.11.0
dataclasses-json==0.5.7
ecdsa==0.18.0
exceptiongroup==1.1.0
executing==1.2.0
fastapi==0.92.0
filelock==3.9.0
frozenlist==1.3.3
func-timeout==4.3.5
greenlet==1.1.3
gunicorn==20.1.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.0.0
iniconfig==2.0.0
jedi==0.18.2
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.2.0
jsonpickle==3.0.1
kaitaistruct==0.10
kiwisolver==1.4.4
kombu==5.2.4
loguru==0.6.0
lxml==4.9.2
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.0
numpy==1.24.2
openai==0.27.0
orjson==3.8.7
outcome==1.2.0
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
playwright==1.27.1
plotly==5.13.1
pluggy==1.0.0
protobuf==3.20.1
pyasn1==0.4.8
pycparser==2.21
pycryptodome==3.17
pycryptodomex==3.17
pydantic==1.10.5
pydub==0.25.1
pyee==8.1.0
Pygments==2.14.0
PyMySQL==1.0.2
pyOpenSSL==23.0.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.2.2
pytest-base-url==2.0.0
pytest-playwright==0.3.1
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
python-slugify==6.1.2
pytz==2022.7.1
pyvis==0.3.2
PyYAML==6.0
redis==4.5.1
regex==2022.10.31
requests==2.28.2
rsa==4.9
schedule==1.1.0
scikit-learn==1.2.2
scipy==1.10.1
selenium==4.8.2
selenium-wire==5.1.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a32
sqlmodel==0.0.8
stack-data==0.6.2
starlette==0.25.0
tenacity==8.2.2
text-unidecode==1.3
threadpoolctl==3.1.0
tiktoken==0.3.0
tomli==2.0.1
tqdm==4.64.1
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.9.2
typing-inspect==0.8.0
typing_extensions==4.5.0
undetected-playwright==0.0.5
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
websocket-client==1.5.1
wsproto==1.2.0
yarl==1.8.2
zipp==3.15.0
zope.event==4.6
zope.interface==5.5.2
zstandard==0.20.0
